build: Remove //build/win:default_exe_manifest dep from nasm target
It adds many configs since `default_exe_manifest` is `source_set`.
https://source.chromium.org/chromium/chromium/src/+/main:build/config/win/manifest.gni;l=98;drc=73a09d1a67a9bab66eba08b51372f431ab709b2d

This is problematic when sanitizer is enabled, because nasm.exe depends
on `clang_rt.asan_dynamic-x86_64.dll` unnecessarily.
https://chromium-review.googlesource.com/c/chromium/deps/nasm/+/6472454/comments/b584d0af_f3818141
(See also https://crrev.com/c/6472454.)

The dependency exists since nasm was added. However, it's not clear if it's necessary. This CL tries to remove it.
https://chromium.googlesource.com/chromium/deps/nasm/+/9596cc0f7bf15a421347556407f9b7a5197e911f

Bug: 411574307
Change-Id: I88c1bf04a9c3f5ce3f8033637448a748615be14b